+++ app/Http/Controllers/Journal/OperationController.php
@@ -5,12 +5,15 @@ namespace App\Http\Controllers\Journal;
 use App\Http\Controllers\Controller;
 use App\Models\Operation;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class OperationController extends Controller
 {
     public function index(Request $request)
     {
+        Gate::authorize(Operation::getPermissions()['operation-list']);
+
         $query = Operation::query();
 
         if ($search = $request->search) {
@@ -24,6 +27,8 @@ class OperationController extends Controller
 
     public function show(Request $request, Operation $operation)
     {
+        Gate::authorize(Operation::getPermissions()['operation-list']);
+
         $query = $operation->accounts()->where('type', \App\Models\Addrbook::TYPE_ACCOUNT);
 
         if ($search = $request->search) {
@@ -38,6 +43,8 @@ class OperationController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize(Operation::getPermissions()['operation-create']);
+
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'nullable|string',
@@ -50,6 +57,8 @@ class OperationController extends Controller
 
     public function update(Request $request, Operation $operation)
     {
+        Gate::authorize(Operation::getPermissions()['operation-edit']);
+
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'nullable|string',
@@ -62,6 +71,8 @@ class OperationController extends Controller
 
     public function destroy(Operation $operation)
     {
+        Gate::authorize(Operation::getPermissions()['operation-delete']);
+
         $operation->delete();
 
         return redirect()->back()->with('success', 'Operation deleted successfully.');
